# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ku-helper backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/me/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "내 알림 목록",
        "description": "현재 사용자가 등록한 모든 알림 설정을 반환한다. 응답 배열의 각 항목은 `type` discriminator 로 TRANSIT / LUNCH / LIBRARY 중 하나의 스키마를 따른다.",
        "operationId": "list_notifications_api_v1_me_notifications_get",
        "responses": {
          "200": {
            "description": "사용자가 보유한 모든 알림 설정 배열",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/_TransitRead"
                      },
                      {
                        "$ref": "#/components/schemas/_LunchRead"
                      },
                      {
                        "$ref": "#/components/schemas/_LibraryRead"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "type",
                      "mapping": {
                        "TRANSIT": "#/components/schemas/_TransitRead",
                        "LUNCH": "#/components/schemas/_LunchRead",
                        "LIBRARY": "#/components/schemas/_LibraryRead"
                      }
                    }
                  },
                  "type": "array",
                  "title": "Response List Notifications Api V1 Me Notifications Get"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "알림 설정 생성",
        "description": "새 알림 설정을 등록한다. 본문은 `type` 으로 분기되는 discriminated union (`TRANSIT` / `LUNCH` / `LIBRARY`). 동일 type 의 다중 등록을 허용한다 — 예: 출근·퇴근용 교통 알림 2 개. config 스키마는 type 별로 다르며 Swagger 의 oneOf 로 노출된다.",
        "operationId": "create_notification_api_v1_me_notifications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/_TransitCreate"
                  },
                  {
                    "$ref": "#/components/schemas/_LunchCreate"
                  },
                  {
                    "$ref": "#/components/schemas/_LibraryCreate"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "생성된 알림 설정 (id, timestamps 포함)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/_TransitRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LunchRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LibraryRead"
                    }
                  ],
                  "title": "Response Create Notification Api V1 Me Notifications Post",
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "TRANSIT": "#/components/schemas/_TransitRead",
                      "LUNCH": "#/components/schemas/_LunchRead",
                      "LIBRARY": "#/components/schemas/_LibraryRead"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          },
          "422": {
            "description": "요청 본문 또는 config 스키마 검증 실패"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/notifications/transit": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "교통 알림 목록만",
        "description": "F-06 단발 도착·F-07 정기 간격 교통 알림(TRANSIT)만 반환한다.",
        "operationId": "list_transit_notifications_api_v1_me_notifications_transit_get",
        "responses": {
          "200": {
            "description": "사용자의 교통 알림 설정 배열",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/_TransitRead"
                  },
                  "type": "array",
                  "title": "Response List Transit Notifications Api V1 Me Notifications Transit Get"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/notifications/lunch": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "점심 알림 목록만",
        "description": "F-09~F-12 점심 알림(LUNCH)만 반환한다.",
        "operationId": "list_lunch_notifications_api_v1_me_notifications_lunch_get",
        "responses": {
          "200": {
            "description": "사용자의 점심 알림 설정 배열",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/_LunchRead"
                  },
                  "type": "array",
                  "title": "Response List Lunch Notifications Api V1 Me Notifications Lunch Get"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/notifications/library": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "도서관 알림 목록만",
        "description": "F-13~F-15 도서관 좌석 알림(LIBRARY)만 반환한다.",
        "operationId": "list_library_notifications_api_v1_me_notifications_library_get",
        "responses": {
          "200": {
            "description": "사용자의 도서관 알림 설정 배열",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/_LibraryRead"
                  },
                  "type": "array",
                  "title": "Response List Library Notifications Api V1 Me Notifications Library Get"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/notifications/{notification_id}": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "단일 알림 조회",
        "description": "지정한 알림 설정 1 건을 반환한다. 본인 소유가 아니면 403, 존재하지 않으면 404.",
        "operationId": "get_notification_api_v1_me_notifications__notification_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "단일 알림 설정",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/_TransitRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LunchRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LibraryRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "TRANSIT": "#/components/schemas/_TransitRead",
                      "LUNCH": "#/components/schemas/_LunchRead",
                      "LIBRARY": "#/components/schemas/_LibraryRead"
                    }
                  },
                  "title": "Response Get Notification Api V1 Me Notifications  Notification Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          },
          "403": {
            "description": "본인 소유가 아닌 알림 설정에 접근 (NOTIFICATION_FORBIDDEN)"
          },
          "404": {
            "description": "지정한 알림 설정이 존재하지 않음 (NOTIFICATION_NOT_FOUND)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "알림 삭제",
        "description": "지정한 알림 설정 1 건을 삭제한다. 본인 소유가 아니면 403, 존재하지 않으면 404. 발송 이력(`notification_history`) 도 함께 정리된다 (cascade).",
        "operationId": "delete_notification_api_v1_me_notifications__notification_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료. 본문 없음"
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          },
          "403": {
            "description": "본인 소유가 아닌 알림 설정에 접근 (NOTIFICATION_FORBIDDEN)"
          },
          "404": {
            "description": "지정한 알림 설정이 존재하지 않음 (NOTIFICATION_NOT_FOUND)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/notifications/transit/{notification_id}": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "교통 알림 수정",
        "description": "교통 알림 설정을 부분 수정한다. `enabled` 만 변경하거나, `config` 를 같이 보내면 **기존 config 를 통째로 대체** 한다(부분 merge X). type 변경은 지원하지 않는다 — 다른 type 으로 바꾸려면 삭제 후 재생성.",
        "operationId": "update_transit_notification_api_v1_me_notifications_transit__notification_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Notification Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 교통 알림 설정",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/_TransitRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LunchRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LibraryRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "TRANSIT": "#/components/schemas/_TransitRead",
                      "LUNCH": "#/components/schemas/_LunchRead",
                      "LIBRARY": "#/components/schemas/_LibraryRead"
                    }
                  },
                  "title": "Response Update Transit Notification Api V1 Me Notifications Transit  Notification Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          },
          "403": {
            "description": "본인 소유가 아닌 알림 설정에 접근 (NOTIFICATION_FORBIDDEN)"
          },
          "404": {
            "description": "지정한 알림 설정이 존재하지 않음 (NOTIFICATION_NOT_FOUND)"
          },
          "422": {
            "description": "요청 본문 또는 config 스키마 검증 실패"
          }
        }
      }
    },
    "/api/v1/me/notifications/lunch/{notification_id}": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "점심 알림 수정",
        "description": "점심 알림 설정을 부분 수정한다. `config` 는 전체 교체 (부분 merge X).",
        "operationId": "update_lunch_notification_api_v1_me_notifications_lunch__notification_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Notification Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LunchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 점심 알림 설정",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/_TransitRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LunchRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LibraryRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "TRANSIT": "#/components/schemas/_TransitRead",
                      "LUNCH": "#/components/schemas/_LunchRead",
                      "LIBRARY": "#/components/schemas/_LibraryRead"
                    }
                  },
                  "title": "Response Update Lunch Notification Api V1 Me Notifications Lunch  Notification Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          },
          "403": {
            "description": "본인 소유가 아닌 알림 설정에 접근 (NOTIFICATION_FORBIDDEN)"
          },
          "404": {
            "description": "지정한 알림 설정이 존재하지 않음 (NOTIFICATION_NOT_FOUND)"
          },
          "422": {
            "description": "요청 본문 또는 config 스키마 검증 실패"
          }
        }
      }
    },
    "/api/v1/me/notifications/library/{notification_id}": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "도서관 알림 수정",
        "description": "도서관 알림 설정을 부분 수정한다. `config` 는 전체 교체.",
        "operationId": "update_library_notification_api_v1_me_notifications_library__notification_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Notification Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 도서관 알림 설정",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/_TransitRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LunchRead"
                    },
                    {
                      "$ref": "#/components/schemas/_LibraryRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "TRANSIT": "#/components/schemas/_TransitRead",
                      "LUNCH": "#/components/schemas/_LunchRead",
                      "LIBRARY": "#/components/schemas/_LibraryRead"
                    }
                  },
                  "title": "Response Update Library Notification Api V1 Me Notifications Library  Notification Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "JWT 누락·만료·서명오류 또는 USER_DELETED"
          },
          "403": {
            "description": "본인 소유가 아닌 알림 설정에 접근 (NOTIFICATION_FORBIDDEN)"
          },
          "404": {
            "description": "지정한 알림 설정이 존재하지 않음 (NOTIFICATION_NOT_FOUND)"
          },
          "422": {
            "description": "요청 본문 또는 config 스키마 검증 실패"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "_LibraryRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "알림 설정 PK.",
            "examples": [
              13
            ]
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "소유자 user.id.",
            "examples": [
              42
            ]
          },
          "type": {
            "type": "string",
            "const": "LIBRARY",
            "title": "Type",
            "description": "알림 분류. 항상 'LIBRARY'."
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "활성화 여부."
          },
          "config": {
            "$ref": "#/components/schemas/LibraryConfig"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "최초 생성 시각."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "마지막 갱신 시각."
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "type",
          "enabled",
          "config",
          "created_at",
          "updated_at"
        ],
        "title": "_LibraryRead",
        "description": "도서관 알림 응답 본문.",
        "examples": [
          {
            "config": {
              "reading_room_id": "301",
              "threshold": 20,
              "urgent_threshold": 5
            },
            "created_at": "2026-05-20T03:23:00+00:00",
            "enabled": true,
            "id": 13,
            "type": "LIBRARY",
            "updated_at": "2026-05-20T03:23:00+00:00",
            "user_id": 42
          }
        ]
      },
...
      "_LunchRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "알림 설정 PK.",
            "examples": [
              12
            ]
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "소유자 user.id.",
            "examples": [
              42
            ]
          },
          "type": {
            "type": "string",
            "const": "LUNCH",
            "title": "Type",
            "description": "알림 분류. 항상 'LUNCH'."
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "활성화 여부."
          },
          "config": {
            "$ref": "#/components/schemas/LunchConfig"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "최초 생성 시각."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "마지막 갱신 시각."
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "type",
          "enabled",
          "config",
          "created_at",
          "updated_at"
        ],
        "title": "_LunchRead",
        "description": "점심 알림 응답 본문.",
        "examples": [
          {
            "config": {
              "highlight_today_pick": true,
              "max_price": 12000,
              "notify_at": "11:00:00",
              "recommend_count": 3
            },
            "created_at": "2026-05-20T03:22:00+00:00",
            "enabled": true,
            "id": 12,
            "type": "LUNCH",
            "updated_at": "2026-05-20T03:22:00+00:00",
            "user_id": 42
          }
        ]
      },
...
      "_TransitCreate": {
        "properties": {
          "type": {
            "type": "string",
            "const": "TRANSIT",
            "title": "Type",
            "description": "알림 분류 discriminator. 항상 'TRANSIT'."
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "생성 직후 활성화 여부. false 면 발송되지 않음.",
            "default": true
          },
          "config": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/_TransitArrival"
              },
              {
                "$ref": "#/components/schemas/_TransitRecurring"
              }
            ],
            "title": "Config",
            "discriminator": {
              "propertyName": "mode",
              "mapping": {
                "arrival": "#/components/schemas/_TransitArrival",
                "recurring": "#/components/schemas/_TransitRecurring"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "type",
          "config"
        ],
        "title": "_TransitCreate",
        "description": "교통 알림 생성 본문. `config.mode` 로 단발/정기를 다시 분기.",
        "examples": [
          {
            "config": {
              "include_congestion": true,
              "line": "4호선",
              "minutes_before": 10,
              "mode": "arrival",
              "station_name": "성신여대입구"
            },
            "enabled": true,
            "type": "TRANSIT"
          }
        ]
      },
      "_TransitRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "알림 설정 PK.",
            "examples": [
              11
            ]
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "소유자 user.id.",
            "examples": [
              42
            ]
          },
          "type": {
            "type": "string",
            "const": "TRANSIT",
            "title": "Type",
            "description": "알림 분류. 항상 'TRANSIT'."
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "활성화 여부."
          },
          "config": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/_TransitArrival"
              },
              {
                "$ref": "#/components/schemas/_TransitRecurring"
              }
            ],
            "title": "Config",
            "discriminator": {
              "propertyName": "mode",
              "mapping": {
                "arrival": "#/components/schemas/_TransitArrival",
                "recurring": "#/components/schemas/_TransitRecurring"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "최초 생성 시각 (ISO 8601, UTC)."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "마지막 갱신 시각."
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "type",
          "enabled",
          "config",
          "created_at",
          "updated_at"
        ],
        "title": "_TransitRead",
        "description": "교통 알림 응답 본문.",
        "examples": [
          {
            "config": {
              "include_congestion": true,
              "line": "4호선",
              "minutes_before": 10,
              "mode": "arrival",
              "station_name": "성신여대입구"
            },
            "created_at": "2026-05-20T03:21:09+00:00",
            "enabled": true,
            "id": 11,
            "type": "TRANSIT",
            "updated_at": "2026-05-20T03:21:09+00:00",
            "user_id": 42
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}